remove putSupplier from endpoints
